requirements.txt
pyyaml
tqdm
numpy
torch
rasterio
